separate bezier p extract WIP
